Replace unsafe fields access in AuthSocialWebViewFragment and refactor webViewClient
